Correct stupid typo in STR2REAL.
Former-commit-id: e1a3829af3f25e9fff681d3698d452ea4120642c

+++ architecture/faust/dsp/proxy-dsp.h
@@ -50,9 +50,7 @@
 #define snprintf _snprintf
 #endif
 
-//#define STR2REAL(x) ((sizeof(FAUSTFLOAT) == 4) ? strtof((x), NULL) : strtod((x), NULL))
-
-inline FAUSTFLOAT STR2REAL(const std::string& x)  { std::stringstream s; s << x; return atoi(s.str().c_str()); }
+inline FAUSTFLOAT STR2REAL(const std::string& x)  { std::stringstream s; s << x; return atof(s.str().c_str()); }
 
 //-------------------------------------------------------------------
 //  Decode a dsp JSON description and implement 'buildUserInterface'
